requirements.txt
scikit-learn
pandas
numpy

apache-airflow~=2.7.3